+++ src/struct/db/storeitem.ts
@@ -1,4 +1,4 @@
-import {prop, DocumentType, index} from '@typegoose/typegoose';
+import {DocumentType, index, prop} from '@typegoose/typegoose';
 import {ReturnModelType} from '@typegoose/typegoose/lib/types';
 
 @index({storeID: 1})
