+++ src/webserver/routes/api/store.ts
@@ -1,21 +1,23 @@
-import express from 'express';
-import sharp from 'sharp';
-import {FilterQuery} from 'mongoose';
-import fs from 'fs';
-import path from 'path';
 import crypto from 'crypto';
+import express from 'express';
+import fs from 'fs';
+import {FilterQuery} from 'mongoose';
+import path from 'path';
+import sharp from 'sharp';
+
 import {ClassroomClient} from '../../../helpers/classroom';
 import {numberFromData, request, Validators} from '../../../helpers/request';
 import {
-	User,
-	IUser,
-	Transaction,
-	Store,
-	IStore,
-	StoreItem,
 	DBError,
+	IStore,
+	IUser,
+	Store,
+	StoreItem,
+	Transaction,
+	User,
 } from '../../../struct';
 import {requestHasUser} from '../../../types';
+
 const router = express.Router();
 
 async function getStorePerms(
